refactor: 세션 관련 테스트 코드를 별도의 패키지에서 관리
