+++ src/test/java/db/SessionDatabaseTest.java
@@ -1,10 +1,9 @@
 package db;
 
-import db.SessionDatabase;
 import model.User;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import webserver.Session;
+import webserver.session.Session;
 
 import java.lang.reflect.Field;
