+++ src/test/java/webserver/session/SessionTest.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.session;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
